cryptsetup-initramfs: retrieve disk encryption key
Use the SEV-SNP attestation flow to securely retrieve the VM
disk encryption key.

This involves preparing an initrd with the tools necessary to
retrieve the SNP attestation report from the SEV firmware and send
the report to an attestation server under the guest owner's control.
The attestation server will reply with the disk encryption key only
if the contents of the attestation report match expected values and
the signature is valid. The disk encryption key is used to unlock
the disk, and the guest is allowed to continue booting.

The entire attestation flow takes place in encrypted memory, so
there is no risk of exposing the disk encryption key outside of
the VM.

Signed-off-by: Jesse Larrew <jesse.larrew@amd.com>