cryptsetup-initramfs: encrypt disk key
The disk encryption key is secret and must be protected during
transit. To accomplish this, the guest will generate a keypair
and include the fingerprint (hash) of the public key in the
attestation report. Then, the public key and the attestation
report are sent together to the guest owner.

The attestation report proves that the guest that generated the
report also owns the public key sent with the report. After
validating the report, the guest owner can use the public key to
encrypt the disk key before sending it to the guest.

Finally, the guest can use the private key to decrypt the disk key.

In addition to assuring confidentiality of the disk key, this also
mitigates MITM attacks where an attacker captures the attestation
report and replays it to the guest owner to try to expose the key.

Signed-off-by: Jesse Larrew <jesse.larrew@amd.com>