nginx: add cgi script to return disk encryption keys
Validate a SEV-SNP guest report and respond with the disk encryption key.

This script requires the directory /var/www/keys to exist and contain files
named after the ID key hash of each guest. The content of the files must be
the plaintext disk encryption passphrase without any trailing '\n' character.

e.g.:
$ echo -n passphrase > /var/www/keys/000000111111222222333333444444
  555555666666777777888888999999aaaaaabbbbbbccccccddddddeeeeeeffffff

where 000000111111222222...ffffff is the SHA384 hash of the guest ID key.

To protect the secrecy of the keys, the directory /var/www/keys should be
readable only by the www-data user. This script will encrypt the passphrase
before sending it in the CGI response.

Signed-off-by: Jesse Larrew <jesse.larrew@amd.com>