identity: change default failure code
Functions are currently written with a "fail closed" pattern that
initializes the return code to EXIT_FAILURE. However, the actual
value of EXIT_FAILURE is the same as EPERM, which results in a
confusing error message about insufficient permissions when this
code is used with perror().

This patch initializes all return codes to -EXIT_FAILURE. As a
result, perror() produces a message about an unknown error code,
which is more aligned with the actual situation (i.e. an unknown
error condition.)

Signed-off-by: Jesse Larrew <jesse.larrew@amd.com>