nginx: simplify http.conf by removing ssl
Each of the examples is secure even without SSL, so turn this off
to simplify the web server configuration.

Note that the web server IP address has been replaced with a
place holder: LISTEN_ADDRESS. This place holder must be replaced
with the actual IP address of the web server before this configuration
will work.

Signed-off-by: Jesse Larrew <jesse.larrew@amd.com>